refactor(gridsat): rename netcdf_geodetic module to geodetic
