+++ src/goesdl/gridsat/__init__.py
@@ -9,7 +9,7 @@ Classes:
 from .dataset_info import GSDatasetInfo
 from .locator_b1 import GridSatProductLocatorB1
 from .locator_gc import GridSatProductLocatorGC
-from .netcdf_geodetic import GSLatLonGrid
+from .geodetic import GSLatLonGrid
 from .image import GSImage
 from .netcdf_info import (
     GSGeospatialInfo,
