+++ src/goesdl/gridsat/geodetic.py
