+++ src/goesdl/gridsat/time.py
@@ -8,7 +8,7 @@ from ..geodesy import RectangularRegion
 from ..netcdf import DatasetView, HasStrHelp, attribute, scalar, variable
 from ..protocols.geodetic import IndexRange
 from ..utils.array import ArrayBool, ArrayFloat64, ArrayInt8, MaskedFloat32
-from .netcdf_geodetic import GSLatLonGrid
+from .geodetic import GSLatLonGrid
 from .metadata import DeltaTimeMetadata, TimeMetadata, VariableMetadata
 
 SECONDS_IN_DAY = 86400
